move generic_battery to obsolete
