+++ obsolete/generic_battery.py
